Fix newsletter service queueing
