+++ CollAction.Tests/MailChimpManagerTests.cs
@@ -2,67 +2,70 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Configuration;
 using CollAction.Services;
+using Microsoft.Extensions.Options;
+using Microsoft.Extensions.Logging;
 
 namespace CollAction.Tests
 {
     [TestClass]
-    public class MailChimpManagerTests
+    public class NewsletterSubscriptionServiceTests
     {
         private string newsletterTestListId = ""; // The ID for our "Test List" list on MailChimp.
 
-        private MailChimpManager BuildMailChimpManager()
+        private NewsletterSubscriptionService BuildNewsletterSubscriptionService()
         {
             IConfiguration configuration = new ConfigurationBuilder().AddUserSecrets<Startup>().Build();
-            newsletterTestListId = configuration["MailChimpTestListId"];
-            MailChimpManager manager = new MailChimpManager(configuration["MailChimpKey"]);
+            var options = new NewsletterSubscriptionServiceOptions();
+            configuration.Bind(options);
+            NewsletterSubscriptionService manager = new NewsletterSubscriptionService(new OptionsWrapper<NewsletterSubscriptionServiceOptions>(options), new LoggerFactory().CreateLogger<NewsletterSubscriptionService>());
             return manager;
         }
 
         [TestMethod]
         public async Task TestGetListMemberStatusOnNonExistentMember()
         {
-            MailChimpManager manager = BuildMailChimpManager();
+            NewsletterSubscriptionService manager = BuildNewsletterSubscriptionService();
             string email = "non-existing-collaction-test-email@outlook.com";
 
-            MailChimpManager.SubscriptionStatus status = await manager.GetListMemberStatusAsync(newsletterTestListId, email);
-            Assert.IsTrue(status == MailChimpManager.SubscriptionStatus.NotFound);
+            NewsletterSubscriptionService.SubscriptionStatus status = await manager.GetListMemberStatusAsync(newsletterTestListId, email);
+            Assert.IsTrue(status == NewsletterSubscriptionService.SubscriptionStatus.NotFound);
         }
 
         [TestMethod]
         public async Task TestAddListMemberAsSubscribed()
         {
-            MailChimpManager manager = BuildMailChimpManager();
+            NewsletterSubscriptionService manager = BuildNewsletterSubscriptionService();
             string email = "collaction-test-email@outlook.com";
 
             await manager.AddOrUpdateListMemberAsync(newsletterTestListId, email, false);
-            MailChimpManager.SubscriptionStatus status = await manager.GetListMemberStatusAsync(newsletterTestListId, email);
-            Assert.IsTrue(status == MailChimpManager.SubscriptionStatus.Subscribed);
+            NewsletterSubscriptionService.SubscriptionStatus status = await manager.GetListMemberStatusAsync(newsletterTestListId, email);
+            Assert.IsTrue(status == NewsletterSubscriptionService.SubscriptionStatus.Subscribed);
         }
 
         [TestMethod]
         public async Task TestDeleteExistingListMember()
         {
-            MailChimpManager manager = BuildMailChimpManager();
+            NewsletterSubscriptionService manager = BuildNewsletterSubscriptionService();
             string email = "collaction-test-email@outlook.com";
 
             await manager.AddOrUpdateListMemberAsync(newsletterTestListId, email, false);
-            MailChimpManager.SubscriptionStatus status = await manager.GetListMemberStatusAsync(newsletterTestListId, email);
-            Assert.IsTrue(status == MailChimpManager.SubscriptionStatus.Subscribed);
+            NewsletterSubscriptionService.SubscriptionStatus status = await manager.GetListMemberStatusAsync(newsletterTestListId, email);
+            Assert.IsTrue(status == NewsletterSubscriptionService.SubscriptionStatus.Subscribed);
 
             await manager.DeleteListMemberAsync(newsletterTestListId, email);
             status = await manager.GetListMemberStatusAsync(newsletterTestListId, email);
-            Assert.IsTrue(status == MailChimpManager.SubscriptionStatus.NotFound);
+            Assert.IsTrue(status == NewsletterSubscriptionService.SubscriptionStatus.NotFound);
         }
 
         [TestMethod]
         public async Task TestDeleteNonExistingListMember()
         {
-            MailChimpManager manager = BuildMailChimpManager();
+            NewsletterSubscriptionService manager = BuildNewsletterSubscriptionService();
             string email = "non-existing-collaction-test-email@outlook.com";
 
             await manager.DeleteListMemberAsync(newsletterTestListId, email);
-            MailChimpManager.SubscriptionStatus status = await manager.GetListMemberStatusAsync(newsletterTestListId, email);
-            Assert.IsTrue(status == MailChimpManager.SubscriptionStatus.NotFound);
+            NewsletterSubscriptionService.SubscriptionStatus status = await manager.GetListMemberStatusAsync(newsletterTestListId, email);
+            Assert.IsTrue(status == NewsletterSubscriptionService.SubscriptionStatus.NotFound);
         }
     }
 }
