+++ CollAction/Controllers/ManageController.cs
@@ -111,9 +111,9 @@ namespace CollAction.Controllers
             {
                 _newsletterSubscriptionService.SetSubscription(user.Email, model.NewsletterSubscription, false);
                 if (model.NewsletterSubscription)
-                    StatusMessage = _localizer["U bent nu ingeschreven voor de nieuwsbrief"];
+                    StatusMessage = _localizer["U bent nu ingeschreven voor de nieuwsbrief (het kan een paar seconden duren voordat u daadwerkelijk uitgeschreven bent)"];
                 else
-                    StatusMessage = _localizer["U bent niet meer ingeschreven voor de nieuwsbrief"];
+                    StatusMessage = _localizer["U bent niet meer ingeschreven voor de nieuwsbrief (het kan een paar seconden duren voordat u daadwerkelijk uitgeschreven bent)"];
             }
             return RedirectToAction(nameof(Index));
         }
